+++ src/bytestream.h
@@ -36,8 +36,8 @@ public:
 	ByteStream(void);
 	virtual ~ByteStream(void);
 
-	int		read(char *buffer, int size);
-	void	write(const char *data, int size);
+	int		read(char *buffer, size_t size);
+	void	write(const char *data, size_t size);
 
 	bool	readBinary(ByteStream &s);
 	bool	readBinary(Serializable &s);
@@ -87,8 +87,8 @@ public:
 	virtual void ignore(int n = 1);
 	
 protected:
-	virtual int readData(char *buffer, int size) = 0;
-	virtual void writeData(const char *data, int size) = 0;
+	virtual int readData(char *buffer, size_t size) = 0;
+	virtual void writeData(const char *data, size_t size) = 0;
 
 private:
 	void	fixEndianess16(char *data);
