+++ src/pipe.h
@@ -39,8 +39,8 @@ public:
 	void close(void);
 
 protected:
-	int readData(char *buffer, int size);
-	void writeData(const char *data, int size);
+	int readData(char *buffer, size_t size);
+	void writeData(const char *data, size_t size);
 
 private:
 	ByteString mBuffer;
