+++ src/socket.h
@@ -45,8 +45,8 @@ public:
 	void close(void);
 
 	// Stream, ByteStream
-	int readData(char *buffer, int size);
-	void writeData(const char *data, int size);
+	int readData(char *buffer, size_t size);
+	void writeData(const char *data, size_t size);
 
 private:
 	socket_t mSock;
